### 2021-06-25 #3412 @kyleniemeyer
Hi @ljakoster, thanks for your interest in JOSS!

Unfortunately, we have extremely limited editor availability to handle submissions right now, and so we need to put your submission on our waitlist untill an appropriate editor becomes available. Thanks for your patience!

It looks like the YAML header in your paper is causing some compilation issues. I *think* the issue is the tabs, rather than spaces, before the items in the list (`  - ` rather than `	-`).
