+++ paper/paper.bib
@@ -11,19 +11,17 @@
 	doi = {10.1016/j.xcrp.2021.100346}
 }
 
-@article {ExcitonDiffusion2020,
-	Title = {Long-range exciton diffusion in molecular non-fullerene acceptors},
-	Author = {Firdaus, Yuliar and Le Corre, Vincent M and Karuthedath, Safakath and Liu, Wenlan and Markina, Anastasia and Huang, Wentao and Chattopadhyay, Shirsopratim and Nahid, Masrur Morshed and Nugraha, Mohamad I and Lin, Yuanbao and Seitkhan, Akmaral and Basu, Aniruddha and Zhang, Weimin and McCulloch, Iain and Ade, Harald and Labram, John and Laquai, Frédéric and Andrienko, Denis and Koster, L Jan Anton and Anthopoulos, Thomas D},
-	DOI = {10.1038/s41467-020-19029-9},
-	Number = {1},
-	Volume = {11},
-	Month = {October},
-	Year = {2020},
-	Journal = {Nature communications},
-	ISSN = {2041-1723},
-	Pages = {5220},
-	Abstract = {The short exciton diffusion length associated with most classical organic semiconductors used in organic photovoltaics (5-20 nm) imposes severe limits on the maximum size of the donor and acceptor domains within the photoactive layer of the cell. Identifying materials that are able to transport excitons over longer distances can help advancing our understanding and lead to solar cells with higher efficiency. Here, we measure the exciton diffusion length in a wide range of nonfullerene acceptor molecules using two different experimental techniques based on photocurrent and ultrafast spectroscopy measurements. The acceptors exhibit balanced ambipolar charge transport and surprisingly long exciton diffusion lengths in the range of 20 to 47 nm. With the aid of quantum-chemical calculations, we are able to rationalize the exciton dynamics and draw basic chemical design rules, particularly on the importance of the end-group substituent on the crystal packing of nonfullerene acceptors.},
-	URL = {https://europepmc.org/articles/PMC7562871},
+@article{SpaceCharge2021,
+author = {Le Corre, Vincent M. and Duijnstee, Elisabeth A. and El Tambouli, Omar and Ball, James M. and Snaith, Henry J. and Lim, Jongchul and Koster, L. Jan Anton},
+title = {Revealing Charge Carrier Mobility and Defect Densities in Metal Halide Perovskites via Space-Charge-Limited Current Measurements},
+journal = {ACS Energy Letters},
+volume = {6},
+number = {3},
+pages = {1087-1094},
+year = {2021},
+doi = {10.1021/acsenergylett.0c02599},
+URL = {https://doi.org/10.1021/acsenergylett.0c02599},
+eprint = {https://doi.org/10.1021/acsenergylett.0c02599}
 }
 
 @Article{Friso2020,
