### 2021-07-01 8d15368
Bug fix (issue with Gfrac), see docs/change_log.txt
