### 2021-07-19 #3412 @danielskatz
You don't need to pause it for a couple of weeks, though you could and should if it is longer than that - here you could just use the @whedon remind mechanism to send a reminder in 2 weeks, or 18 days, ...
