### 2021-07-19 #3412 @danielskatz
oh, sorry, didn't read enough - yes, you could use paused in this case.
