### 2021-08-10 #3412 @ljakoster
hi @rkurchin <https://github.com/rkurchin>,

sorry for the slow reply, just got back to the office.
May I suggest:
-dawbaron
-eteq
-danehkar
-shahmoradi

best wishes,
Jan-Anton

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Mon, 9 Aug 2021 at 15:30, Rachel Kurchin ***@***.***>
wrote:

> 👋 @ljakoster <https://github.com/ljakoster>, just checking in again,
> whenever you're back from holiday and ready to get this going again, do let
> us know!
>
