### 2021-08-10 #3412 @rkurchin
👋  Hi there @shahmoradi and @danehkar, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Please feel free to let me know if you have any questions about the process.
