### 2021-08-10 #3412 @rkurchin
Thanks. Our list of volunteers is a bit thin on folks who list themselves as knowing any Pascal and also as doing physics, so if you feel like you can tackle the subject matter, I think you'd be a perfectly fine fit. Let me know what you think!
