### 2021-08-10 #3412 @rkurchin
Once we have at least two reviewers, we'll start the review issue (this is the pre-review issue). The review process should be quite interactive, with reviewers filing issues (or even PR's) to address any areas that are lacking, either in the code itself or its documentation. So the duration of that can vary a fair bit depending on how quick that back-and-forth is. But once the review starts, you should definitely not feel as though you have to have gone through every element of the checklist before asking the author questions or filing issues on the repo, e.g. if you have issues installing or running the code.

Anyway, for the moment you can sit tight until we get another reviewer, but hope this helps!
