### 2021-08-23 #3412 @rkurchin
👋 Hi there @UTH-Tuan, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Please feel free to let me know if you have any questions about the process.

I recognize this submission is a bit outside your subject area expertise, so I'm primarily asking you to review the software design side of it, as we don't have too many reviewer volunteers with Pascal expertise. I'll recruit another complementary reviewer (see next message) to focus primarily on the subject matter.
