### 2021-08-23 #3412 @rkurchin
👋 Hi there @dalonsoa, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Please feel free to let me know if you have any questions about the process.

I know you didn't note Pascal expertise in your volunteer signup, so I'm hoping that as a semiconductor/PV expert that you can primarily review the subject matter side of things, and we'll lean on other reviewers (see prior comment) to handle the specifics of software design in Pascal.
