### 2021-09-07 #3412 @MikeHeiber
Sorry for the slow response. I've been on holiday the past several days. Unfortunately, I must declare a conflict of interest to providing an unbiased review. I am a recent collaborator of the author's. I would suggest @roderickmackenzie (roderick.mackenzie@nottingham.ac.uk) who has published related simulation codes in the past. He is also much more of an expert in the details of drift-diffusion simulations than I am and should be able to provide a high quality scientific review.
