### 2021-09-07 #3412 @rkurchin
No problem. Thanks for the straightforwardness and alternative suggestion! @roderickmackenzie, we'd love to have your help here if you're willing! We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
