### 2021-09-13 #3412 @rkurchin
emailed Prof. MacKenzie today to follow up:

> Dear Prof. MacKenzie,
>
>
> I’m an editor for the Journal of Open-Source Software (JOSS). I’m currently seeking reviewers for a manuscript on drift-diffusion simulations (see more details here), and you were suggested as a qualified reviewer.
>
> JOSS is a free, open-source, community driven and developer-friendly online journal (no publisher is seeking to raise revenue from the volunteer labor of researchers!). The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged.
>
> JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html . Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review).
>
> Would you be able to review this submission for JOSS? If not, could you provide recommendations of other qualified reviewers?
>
> Thanks in advance,
> Rachel Kurchin
