### 2021-09-14 #3727 @whedon
[REVIEW]: SIMsalabim: A time resolved drift-diffusion simulation software
**Submitting author:** @ljakoster (<a href="http://orcid.org/0000-0002-6558-5295">Lambert Jan Anton Koster</a>)
**Repository:** <a href="https://github.com/kostergroup/SIMsalabim" target ="_blank">https://github.com/kostergroup/SIMsalabim</a>
**Version:** v4.05
