### 2021-09-15 #3727 @ljakoster
@rkurchin, good point. We have only published brief descriptions of the
code as one does in the method-section of a paper. So, only the
mathematical form of the equations and the iteration loop (simplified). The
difference equations, the code, the parameters, the manual are all new.

jan anton

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Wed, 15 Sept 2021 at 15:19, Rachel Kurchin ***@***.***>
wrote:

> @ljakoster <https://github.com/ljakoster>, I think Rod's comments above
> give a solid starting point for things to address on the technical side. I
> wanted to jump in with a slightly more "administrative" query as well:
>
> Given that this code has been around for a long time and has suggested
> citations in the repo, it's important that we ensure this submission isn't
> in violation of JOSS's self-plagiarism policy. Can you clarify what aspects
> of this submission have *not* been previously published?
>
