### 2021-09-16 #3412 @ljakoster
@roderickmackenzie many thanks for your input; we'll be working on those
and will post a new version (I'm assuming that's how it's done).

jan anton

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Tue, 14 Sept 2021 at 08:48, Roderick MacKenzie ***@***.***>
wrote:

> Hi There!,
> Thanks for your mail. Sorry for not seeing this I have a special e-mail
> account setup for high volume stuff which I don't check.. and this went
> into it.. So as far as I understand the review process, I just write
> comments here and then the authors respond?
>
> I am happy to take a look but will need a few days to compile and poke at
> the code base. I've not used Pascal since the Borland Delphi days of 2000..
>
> I've read the manual and had a quick poke around. Some initial comments
> are:
> *Nice manual.
> *Nice scientific description in the manual.
> *The code seems clear on first impressions.
> *I like the hat logo!
> *I also like your implementation of the bernoulli function, might be
> faster than what I do. Exponentials are always a pain to calculate as they
> are done iteratively... I would be good though if you could point to
> literature or show a graph that what you do works.
> *The model is clearly a nice thing and a benefit to the community. The
> more people are simulating these devices the better as we will have a
> chance of pushing renewable stuff forward...
>
> Here are some general questions to kick off with:
>
>    1.
>
>    Your use of the V in equations 2.1 -2.6 is a bit confusing because
>    commonly V is only used in the DD equations when hetero-junctions are
>    *not* taken into account. When the model has been adjusted to handle
>    hetero-junctions (as yours seems to be 2.13-2.14) it's conmen to write Ec
>    and Ev in the DD equations instead of V, to explicitly say that heterosteps
>    can drive current. This is especially confusing as your fig 2.1 shows
>    hetero steps so made me scratch me head as soon as I started reading until
>    I reached the interfaces section. If you rewrite in terms of Ec and Ev you
>    can get rid of equations 2.13-2.14, it would make the description look more
>    like the classical DD literature as used for things like the III-V
>    materials - and make things a bit clearer.
>    2.
>
>    As I understand it there are two parts to the model a transient part
>    and a steady state parts. In the intro you talk about transient
>    photovoltage and AC analysis. However your equation 2.22 is the steady
>    state form of the SRH equation, so you are assuming your trapped carrier
>    distribution is in quasi-steady state. This means that although you can do
>    time domain simulations your trapped carriers will always think they are in
>    quasi-equilibrium. This won't matter for DC or for slow small perturbation
>    stuff. But you won't for example be able to simulate things far from
>    equilibrium, e.g. ToF transients are an extreme example of this where
>    carriers are initially captured then gradually released from deeper and
>    deeper traps as time goes on. TPV transients could also meet this large
>    signal condition depending on the power of the laser pulse, and AC
>    depending on carrier trapping/detrapping rates and the frequency you are at
>    (and trap depth). CELIV will also fall into large perturbation regime
>    etc... etc... I think the user should be alerted to the fact that your
>    carriers are in SS and the model does not explicitly solve the SRH
>    equations in time domain. Many users won't grasp this so I think it needs
>    to be clearly pointed out and what the implications of this are for results.
>    3.
>
>    I'm assuming that your trapped carriers introduced through equation
>    2.22 don't interact with the field? As for some devices (organics) quite a
>    lot of the charge will be trapped (>90%) so this could affect the results.
>    If the trapped charge does not affect the field I think the user should
>    also be told about this in the manual. Maybe include a limitations section
>    in the manual? Again, it depends on what the user is using the code for. If
>    it's some perovskite device with not many traps then this may not be an
>    issue, but if it's some ultra trappy organic thing then..
>
> I think my main concern with points 2 and 3 is that the user knows what
> he/she is dealing with. There are a tonne of papers where people just use
> software (aka COMSOL etc..) to produce a paper without understand what is
> going on under the hood and results in miss interpretation of data which is
> not a good thing.
>
>    1. Would be nice if you could add some makefiles. Also would be good
>    if you could tell me (the user) which package I need to install on Ubuntu
>    to get pascal. I'm not super excited about installing free pascal binaries
>    from anything but the Ubuntu repos due to security reasons.
>
> I'll install and play with over the next week then get more computer
> related comments back to you..
>
> best,
> Rod
>
