### 2021-09-16 #3727 @UTH-Tuan
I am bit confused by documentation to compile it. Sorry, my previous experience was in Object Pascal using Delphi IDE. So it might be helpful to get more clarification on how to compile it. Specifically I am thrown off by the bracket placeholder ( [zimbt] )... is this supposed to be the path to the ZimbT folder?
