### 2021-09-17 #3727 @ljakoster
We'll improve the description of how to compile the code (and where to get
the compiler re Rod's comments). About ZimT, it should work by 1)
installing the fpc compiler (no IDE is needed)
2) go to the ZimT folder (note: no b in spelling).
3) type fpc zimt.pas

jan anton
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>
