### 2021-10-05 #3727 @ljakoster
@rkurchin nope, I was kind of waiting for the reviews by the other
reviewers so we can deal with them in 1 go. However, we will start working
on Rod's comments. On a more technical note: how do I do this? do I simply
address the issues, put a rebuttal somewhere and upload a new version on
Github?

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Tue, 5 Oct 2021 at 14:54, Rachel Kurchin ***@***.***>
wrote:

> @ljakoster <https://github.com/ljakoster>, any updates on Rod's review?
>
> @UTH-Tuan <https://github.com/UTH-Tuan>, did the response above help you
> to compile the code?
>
> @shahmoradi <https://github.com/shahmoradi>, reminder to please start
> your review!
>
