### 2021-10-05 #3727 @rkurchin
Make any updates to the code and/or manuscript that are needed to address the issues, and describe the changes here (or, if you prefer to open issues in your own repository to track things individually, that's also fine, just mention this issue in them and link to them here). If you update the manuscript, you can refresh the proof here by running `@whedon generate pdf` (note that this must be the first thing in your comment for Whedon to trigger properly).
