+++ README.md
@@ -1,4 +1,4 @@
-# SIMsalabim
+# SIMsalabim Project
 
 SIMsalabim: A 1D drift-diffusion simulator for semiconductor devices (LEDs, solar cells, diodes, organics, perovskites). 
 
@@ -9,15 +9,23 @@ SIMsalabim: A 1D drift-diffusion simulator for semiconductor devices (LEDs, sola
 4. [How to cite](#how-to-cite)
 5. [How to compile](#how-to-compile)
 6. [Pre-compiled binaries](#pre-compiled-binaries)
-7. [Scientific publications based on SIMsalabim](#scientific-publications-based-on-simsalabim)
+7. [Quickstart guide](#Quickstart guide)
+8. [Scientific publications based on SIMsalabim](#scientific-publications-based-on-simsalabim)
 
 ## Introduction
 
-SIMsalabim can be used to simulate current-voltage (JV) characteristics of semiconductor devices. It includes the effects of generation, recombination and trapping of electrons and holes, the effect of ions and dopants, and self-consistently solves the electric field that results from all charged species.
+SIMsalabim and ZimT can be used to simulate current-voltage (JV) characteristics of semiconductor devices. This includes the effects of generation, recombination and trapping of electrons and holes, the effect of ions and dopants, and self-consistently solves the electric field that results from all charged species. The routines that make up this project are used in two different guises: SIMsalabim and ZimT. The former is a steady-state version and is the 'main' code. The latter, ZimT, can be used to simulate transients. 
+The project folder is structured as follows:
+- Binaries: compiling the code from source is easy, but if you do not want to do this then you can try the binaries in this folder
+- Docs: this folder contains a reference manual, developer guidelines, and a log of all changes to the code
+- SIMsalabim: contains the steady-state version of this software
+- Tests: contains a number of tests to assess the functioning of the codes (SIMsalabim and ZimT)
+- Units: this is where the heart of the code resides. These units are shared between SIMsalabim and ZimT
+- ZimT: the transient equivalent of SIMsalabim
 
 ## Copyright and license
 
-SIMsalabim is licensed under the GNU Lesser General Public Licence version. The details of this licence can be found in the files COPYING.txt and COPYING.LESSER.txt. 
+The SIMsalabim project is licensed under the GNU Lesser General Public Licence version 3. The details of this licence can be found in the files COPYING.txt and COPYING.LESSER.txt. 
 Several authors (all from the University of Groningen) have contributed to the code: 
 - Dr T.S. (Tejas) Sherkar
 - Dr V.M. (Vincent) Le Corre
@@ -27,7 +35,7 @@ Several authors (all from the University of Groningen) have contributed to the c
 
 ## Instructional videos
 
-In order to help new users to get the most out of SIMsalabim, we have recorded a few instructional videos:
+In order to help new users to get the most out of SIMsalabim/ZimT, we have recorded a few instructional videos:
 
 1 [Installation and basic use](https://youtu.be/0mtpGJMnbFE "Installation and basic use")
 
@@ -43,41 +51,98 @@ For the use of perovskite solar cells, the following paper offers a detailed des
 
 ## Pre-compiled binaries
 
[…]

[…]
