### 2021-10-20 9d22381
readme update link to binaries in releases
