### 2021-10-20 #3727 @roderickmackenzie
@kostergroup Hi, thanks for your answer. I'm still confused about:
*How the charge for bulk traps in equation 2.1 i.e. C(x) links to equation 2.23.
*How is C(x) is calculated is not in the manual and I think the term needs breaking down term by term, please give a term for trapped electrons, trapped holes, and ions also listing the doping terms. I find C(x) confusing, maybe just spell out the terms in the equation.
*I don't understand this sentence "The density of trapped charges follows from the density of traps and their filling." that seems to contradict equation 2.23, as that is not what SRH does.
*I don't understand what rate equations you use for bulk trapping, where are they in the manual? Where is the capture cross sections, how are escape rates calculated. Equation 2.23 will only work for steady state.
*This sentence should not be true for SRH time domain "When electrons and holes recombine via a bulk trap (SHR recombination in the bulk), then the recombination rates for electrons and holes are equal in every grid point". As each trap should have it's own IMREF so electron and hole rates can be independent (in time domain)
*Can you simulate a ToF transient with an exponential distribution of tail states, place the cell in negative bias then from the resulting current transient re-extract the DoS as I did here: dx.doi.org/10.1021/jp4010828 using equation 1. This should take a couple of hours but would be an acid test of traps in your model.
