### 2021-10-20 #3727 @ljakoster
@roderickmackenzie <https://github.com/roderickmackenzie>
*How the charge for bulk traps in equation 2.1 i.e. C(x) links to equation
2.23.
C(x) is simply the sum of all charged species at x. For bulk traps, their
contribution follows from SRH statistics as described, for example, in  W.
Shockley, W. T. Read, Phys. Rev. 1952, 87, 835–842. Eq. 2.23 only describes
the corresponding recombination rate, so it does not link to C(x).

*How is C(x) is calculated is not in the manual and I think the term needs
breaking down term by term, please give a term for trapped electrons,
trapped holes, and ions also listing the doping terms. I find C(x)
confusing, maybe just spell out the terms in the equation.

The details of how C(x) are calculated are 1) in the code, 2) in ref. 2 in
the reference manual. Please note, that our interface recombination model
is an extension of the SRH work: it puts traps at an interface and charges
can be captured & emitted to the conduction and valence bands on either
side. Therefore, it consists of 8 rates, rather than 4. The resulting
expression (and the way we linearize this in the Poisson solver) warrants a
separate publication (=ref 2 in manual, in prep.). Our treatment of this
process also goes beyond what is incorporated in, e.g., IonMonger as their
interfacial charges do not show up in the Poisson equation and they
simplify the rates.

*I don't understand this sentence "The density of trapped charges follows
from the density of traps and their filling." that seems to contradict
equation 2.23, as that is not what SRH does.

-This sentence expresses that there is a number of traps per volume (the
density of the traps), and they will have a filling. Their product is,
naturally, the density of trapped charges. What we do is exactly what SRH
in their papers do for bulk traps. It is, therefore, quasi-steady-state as
remarked. We are writing a section in the reference manual to point this
out (i.e. the limitations and strengths of the model). Equation 2.23 does
not specify the filling, but rather resulting recombination for bulk traps
in steady-state.

*I don't understand what rate equations you use for bulk trapping, where
are they in the manual? Where is the capture cross
sections, how are escape rates calculated. Equation 2.23 will only work for
steady state.

-Yes, quasi equilibrium / steady-state only, as remarked. Please note, we
only replied to a fraction of your comments (as noted) as we are working on
the rest. The strengths and limitations of our code will be outlined in a
section in the reference manual. Future plans include moving beyond the
quasi steady-state approach.

*This sentence should not be true for SRH time domain "When electrons and
holes recombine via a bulk trap (SHR recombination in the bulk), then the
recombination rates for electrons and holes are equal in every grid point".
As each trap should have it's own IMREF so electron and hole rates can be
independent (in time domain)

-see above.

*Can you simulate a ToF transient with an exponential distribution of tail
states, place the cell in negative bias then from the resulting current
transient re-extract the DoS as I did here: dx.doi.org/10.1021/jp4010828
using equation 1. This should take a couple of hours but would be an acid
test of traps in your model.

-No, this is not possible as we do not have an exponential distribution of
traps. Previous version of SIMsalabim did include such distributions (and a
Gaussian one as well), but we have removed them as they were hardly used by
the users of the code.
A further acid test of traps in our model consists test 5.

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Wed, 20 Oct 2021 at 11:28, Roderick MacKenzie ***@***.***>
wrote:

> @kostergroup <https://github.com/kostergroup> Hi, thanks for your answer.
> I'm still confused about:
> *How the charge for bulk traps in equation 2.1 i.e. C(x) links to equation
> 2.23.
> *How is C(x) is calculated is not in the manual and I think the term needs
> breaking down term by term, please give a term for trapped electrons,
> trapped holes, and ions also listing the doping terms. I find C(x)
> confusing, maybe just spell out the terms in the equation.
> *I don't understand this sentence "The density of trapped charges follows
> from the density of traps and their filling." that seems to contradict
> equation 2.23, as that is not what SRH does.
> *I don't understand what rate equations you use for bulk trapping, where
> are they in the manual? Where is the capture cross sections, how are escape
> rates calculated. Equation 2.23 will only work for steady state.
> *This sentence should not be true for SRH time domain "When electrons and
> holes recombine via a bulk trap (SHR recombination in the bulk), then the
> recombination rates for electrons and holes are equal in every grid point".
> As each trap should have it's own IMREF so electron and hole rates can be
> independent (in time domain)
> *Can you simulate a ToF transient with an exponential distribution of tail
> states, place the cell in negative bias then from the resulting current
> transient re-extract the DoS as I did here: dx.doi.org/10.1021/jp4010828
> using equation 1. This should take a couple of hours but would be an acid
> test of traps in your model.
>
