### 2021-10-25 679c09f
Update README.md
Add section to quickstart outlining what to do for platforms other than Windows or Linux.
