### 2021-11-01 #3727 @UTH-Tuan
@ljakoster if possible, just provide a hyperlink to https://github.com/kostergroup/SIMsalabim/blob/master/Docs/Developer_guidelines.md for the mention of it under Miscellaneous section.
