### 2021-11-02 #3727 @ljakoster
@UTH-Tuan: thanks, that's an excellent suggestion. We're rewriting the
readme file and will include the link.

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>
