+++ SIMsalabim/device_parameters.txt
@@ -1,29 +1,29 @@
 ** SIMsalabim Device data:
 ** Don't change the order of the parameters, comments can be added anywhere,
 ** but only after an '*'. Use '**' if you want your comment to be left-justified.
-** version: 4.20
+** version: 4.22
 
 **General**************************************************************************
 T = 295                  * K, absolute temperature
-L = 200e-9               * m, device length/thickness
+L = 400e-9               * m, device length/thickness
 eps_r = 22               * relative dielectric constant
-CB = 2.0                 * eV, conduction band edge
-VB = 6.0                 * eV, valence band edge
-Nc = 1e26                * m^-3, DOS of conduction and valence bands
+CB = 3.3                 * eV, conduction band edge
+VB = 5.0                 * eV, valence band edge
+Nc = 1e25                * m^-3, DOS of conduction and valence bands
 n_0 = 0                  * m^-3, ionised n-doping
 p_0 = 0                  * m^-3, ionised p-doping
 
 **Mobilities************************************************************************
-mun_0 = 1e-7             * m^2/Vs, zero field mobility
-mup_0 = 1e-7             * m^2/Vs, zero field mobility
+mun_0 = 1e-4             * m^2/Vs, zero field mobility
+mup_0 = 1e-4             * m^2/Vs, zero field mobility
 mob_n_dep = 0            * 0 : const. mob, 1 : field-dependent
 mob_p_dep = 0            * 0 : const. mob, 1 : field-dependent
 gamma_n = 1e-4           * (m/V)^0.5, field dependence of mob, Poole-Frenkel form
 gamma_p = 0              * (m/V)^0.5, field dependence of mob, Poole-Frenkel form
 
 **Contacts**************************************************************************
-W_L = 2.0                * eV, work function left electrode
-W_R = 2.0                * eV, work function right electrode
+W_L = 3.4                * eV, work function left electrode
+W_R = 4.9                * eV, work function right electrode
 Sn_L = -1e-7             * m/s, surface recombination of electrons at the left electrode
 Sp_L = -1e-7             * m/s, surface recombination of holes at the left electrode
 Sn_R = -1e-7             * m/s, surface recombination of electrons at the right electrode
@@ -33,24 +33,24 @@ Rshunt = -1e0            * Ohms m2, shunt resistance. Use negative value for inf
 Rseries = 0e-4           * Ohms m2, series resistance.
 
 **Transport layers******************************************************************
-L_LTL = 0e-9             * m, thickness left TL
-L_RTL = 0e-9             * m, thickness right TL
-Nc_LTL = 1e26            * m^-3, DOS of left TL
-Nc_RTL = 1e26            * m^-3, DOS of right TL
+L_LTL = 30e-9             * m, thickness left TL
+L_RTL = 30e-9             * m, thickness right TL
+Nc_LTL = 1e25            * m^-3, DOS of left TL
+Nc_RTL = 1e25            * m^-3, DOS of right TL
 doping_LTL = -0e24       * m^-3, doping in left TL
                          * if >0 p-type doping if <0 n-type doping
 doping_RTL = 0e24        * m^-3, doping in right TL
                          * if >0 p-type doping if <0 n-type doping
-mob_LTL = 1e-6           * m2/Vs, mobility of left TL, constant (case 0) only.
-mob_RTL = 1.5e-8         * m2/Vs, mobility of right TL, constant (case 0) only.
+mob_LTL = 1e-7           * m2/Vs, mobility of left TL, constant (case 0) only.
+mob_RTL = 1e-7         * m2/Vs, mobility of right TL, constant (case 0) only.
 nu_int_LTL = 1e3         * m/s, interface transfer velocity, left TL
[…]
