### 2021-11-04 #3727 @UTH-Tuan
@ljakoster also this statement from above:

> We did not put a comparison in the paper as we did not want to offend
> anyone working in the field. Most researchers working in this field know of
> one another so we didn’t really feel like point out pros and cons of
> specific software.

I don't think that's a satisfactory response. I think you can mention the other software without comparison (and without "offending" them).
