### 2021-11-05 #3727 @ljakoster
@UTH-Tuan
Yes, we can surely mention them, add a link and a short description of
other software, thanks for the suggestion.
I thought you were asking for a 'table' with pros and cons and features.
All the codes out there have their strengths and uses and I would not feel
comfortable with a table like that, especially since those codes (like
ours) keep on evolving so such a table might become inaccurate/out-dated in
the future. I apologize for the confusion.

best wishes,
Jan-Anton

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>
