### 2021-11-09 #3727 @shahmoradi
@kostergroup Thank you for your corrections and explanations. Given the corrections made, I have checked all boxes in my review. However, I still feel uncomfortable with not having automated tests of some sort in this library. The test results in the repository in PDf format look nice. But will they look so for the next release of the package? Who will know how to generate the test results and make appropriate e(I guess visual) comparisons to verify the accuracy of the output of the software with the next release. This becomes especially a problem given that the code for generating the plots of the test PDF files does not appear to have been shared in the repository. As I stated before, the output of the tests should be (at least semi-) automatically verified and not reply on human intervention and visual checks. The only exception is when the library's development stage is terminal and no future changes are expected to occur.

I am going to make the final decision on this particular item in agreement with other referees' opinions (@roderickmackenzie and @UTH-Tuan).
All other items look fine.

>
> @shahmoradi <https://github.com/shahmoradi> Reviewer Amir Shahmoradi Comments for the authors: - The paper is well-written and the statement of the need is reasonable. - A comparison with other available comparable software to SIMsalabim would be highly desirable. ANSWER We did not put a comparison in the paper as we did not want to offend anyone working in the field. Most researchers working in this field know of one another so we didn’t really feel like point out pros and cons of specific software. To name a few, SCAPS (free-ware, code not shared) does not include ionic movement, IonMonger (open-source) does not include the charge of carriers trapped at interfaces (and has a simplified expression for the SRH recombination process), Setfos (commercial) does include ionic motion but their description of interface recombination (very important in perovskite devices) is crude. All these codes have their strengths as well: Setfos has a GUI (as does SCAPS), SCAPS is much easier to use if working on established solar cells, IonMonger can handle higher ion densities. Behind the scenes we quantitatively compared SIMsalabim (and its transient equivalent ZimT) with SCAPS and, to a lesser extent, Setfos. The agreement is excellent. The comparisons are performed by Vincent Le Corre and they can be accessed via https://github.com/VMLC-PV/Test-SIMsalabim-ZimT. Additionally, to highlight the major features of our project (without offending our colleagues), we will add a section to the readme.md file to highlight what is so nice about our work (in addition, of course, to a section on its limitations). - Licensing issues: - There are multiple CopyRight files in the project in different subfolders. Only one should really exist and if parts of the code are licensed differently, then this should be clearly stated in the front page README file licensing section. But I doubt if this is indeed the case as all license files appear to be either GNU or GNU Lesser. - I think it suffices to have only one set of license files at the root of the project. If the authors are worried about users' lack of attention to the license, I suggest adding the licensing information at the top of each source file. Having too many redundant files in the repository, only adds to the user's confusion. - My suggestion is to rename all licensing files to something that clearly conveys what the files indeed contain, like: LICENSE, LICENSE.txt, LICENSE.md, ... It seems to me that JOSS requires the existence of a file named LICENSE in the root of the project, which is missing in this repository. ANSWER Indeed, all files are licensed under GLPL version 3. It is per GNU recommendations that we call the files ‘COPYING’ and ‘COPYING.LESSER’, see https://www.gnu.org/licenses/gpl-howto.html and we would like to stick with GNU as they wrote the licence. However, the reviewer appears to be right in saying that 1 file is enough. So we have removed all such files except the ones in the root of the project. The licensing information is described in the readme file and in Docs/Reference_Manual.pdf. Every source file points out (at the top of the file) that the project comes with a license file. However, we have not included the license text itself in every source file as that would compromise the readability of the code. - Binary releases: - The binaries provided in the repository work fine for the given sample. - However, it seems odd and highly unusual to release binaries as part of the Git project. I am even surprised by the fact that GitHub allows the uploading of binaries to a Git project hosted on GitHub. I suggest the authors upload the binaries to the release page of their project. There are currently no releases in published under this project. The binaries along with the example files should go there. Authors could separate the builds for different platforms into separate release compressed files. ANSWER We thank the reviewer for this excellent suggestion; we were not aware of github’s releases page. We have moved the binaries to this dedicated page and have updated the readme file accordingly. It works a charm. I know that to the more computer-savvy the inclusion of binary files seems rather odd (and I agree). However, most of the potential users of the code that we have spoken with are daunted by the prospect of having to download the fpc compiler and having to compile the code---no matter how trivial. So it seems there is a place for pre-compiled binaries. - Documentation: - Missing README.md files: - Every folder should ideally contain a README.md file describing the reasons for the existence of the folder and a description of its contents. For example, what is the purpose of zimT folder? Is it a separate project? If it is a separate project, then the authors may want to put it in a separate repository. If it is not a separate project, then its connection with SIMsalabim should be clarified. - User Documentation: The User documentation looks fine on GitHub. But the file seems to be corrupted on Windows OS. The authors should check if the user documentation PDF file opens properly on all supported operating systems. It is possible that this PDF file was generated on a macOS or Linux OS and it is incompatible with Windows. In any case, it needs a fix. - change_log.txt Should likely be moved to the root of the project. - Ideally, the project would also benefit from a developer documentation devoted to guiding the future developers about the structure of the code. Consider this question: If I decide to make any contributions to this project, can I start today independently, fork the project, understand the structure of the project, and make meaningful contributions to the project to merge with the main branch? This is where the developer documentation becomes handy. Right now, this project appears to become an orphan as soon as the primary contributor to the project (kostergroup) stops contributing to it, unless there is a developer guideline for the potential future developers. ANSWER We have updated the readme file in the root of the project to point out how the project is divided over the several subfolders (docs, units, etc); in other words, to explain the purpose of the different folders. We would much rather have a single readme file as that makes it easier to maintain the project. Please note, we are an academic group of relatively small size so this is an important issue for us. We have chosen to leave change_log.txt in the Docs folder: too many files in the root folder would likely detract from the important files such as readme.md and the licensing. To the computer-savvy this may seem like a non-issue, but most of our target audience are (early-stage) PhD students in chemistry and physics who may not realise the importance and meaning of a proper license and things like a readme.md file. The Reference_Manual.pdf file was indeed generated on Linux but renders perfectly on our WIN 10 installations so we are unable to reproduce this problem. As explained in the Reference Manual, ZimT is the transient version of SIMsalabim and is indeed part of this project (most of the code is shared between the two programs). This should have been explained in the readme file and in the paper itself; we apologise for the oversight. We have rewritten the readme file to explain how ZimT and SIMsalabim are related. The overall project is called SIMsalabim and that is also where the main (=steady-state) program SIMsalabim gets its name from. We spent a lot of time over the past 2 years to integrate the two codes as they solve the same system of equations. Having them in 1 project makes it much easier to maintain (as we have found out over the past decade). We followed the reviewer’s suggestion to lay down some developer guidelines: In Docs/developer_guidelines.md we have outlined how the project is structured and how it could be extended. - Testing and Code Coverage: - There exists a folder named Tests in the project. But I could not find any code that performs any tests of the codebase functionality here in a verifiable manner. The tests in this folder resemble more a set of example input files. The authors should consider this question: How could a potential user of this library ensure that the library's functionality is verifiable at any time in the future? Even if the code works fine today and is accurate, how do we know the next person who works on this project will not introduce bugs and errors in the code or cause regression inadvertently? ANSWER The reviewer is right: the next person should be able to verify the outcomes of the code. Rather than relying on comparisons with other software (see above) we rely on physics to test the code. As a physicist, I for one find a graph that compares the numerical result with a known physical model (in a suitable limit) very compelling. JOSS requires that there is some means of testing the code. Ideally this takes the form of automated testing. However, manual tests are also allowed and that is what we have done. The different folders (testx) indeed include the SIMsalabim or ZimT input device_parameters.txt that will generate the data. To run the test all that is required is to copy the parameter files to the SIMsalabim / ZimT folder and run either program. We have modified the file tests.md to make clear how the testing should be done (we have also pointed this out in the developer instructions). At this stage of the project (SIMsalabim started back in 2001) it is too late in the day to introduce automated unit tests for the code. Alas, we don’t have the time to do this. - Compilation: - I have tested the compilation of the two codebases on Windows and it seems to compile and run fine with the supplied example input file.
> […](#)
> -- Prof. Dr. L.J.A. (Jan Anton) Koster Zernike Institute for Advanced Materials University of Groningen open-source simulator for perovskite and organic PV <https://github.com/kostergroup/SIMsalabim> group website <http://www.koster-group.nl>
> On Wed, 20 Oct 2021 at 10:06, Koster, Jan Anton ***@***.***> wrote: @roderickmackenzie <https://github.com/roderickmackenzie> Review Rod MacKenzie ANWER We have addressed most issues raised by this reviewer, but are not yet done with some. Please find below the answers to the issues that we have addressed so far. I've read the manual and had a quick poke around. Some initial comments are: *Nice manual. *Nice scientific description in the manual. *The code seems clear on first impressions. *I like the hat logo! *I also like your implementation of the bernoulli function, might be faster than what I do. Exponentials are always a pain to calculate as they are done iteratively... I would be good though if you could point to literature or show a graph that what you do works. ANSWER We have added a short description of the approximation and a graph showing the resulting error in the reference manual (section 2.1). In doing so, we have also improved our implementation of the approximated Bernoulli function. The overall error is now smaller than 3.5E-4 (absolute), which implies that the relative error is smaller than 8E-4. This can be found in the reference manual section 2.1 and fig 2.2. *The model is clearly a nice thing and a benefit to the community. The more people are simulating these devices the better as we will have a chance of pushing renewable stuff forward... ANSWER We thank the review for their nice comments. We had the logo designed (albeit on the cheap) as a token to our dedication to the software suite. I'm assuming that your trapped carriers introduced through equation 2.22 don't interact with the field? As for some devices (organics) quite a lot of the charge will be trapped (>90%) so this could affect the results. If the trapped charge does not affect the field I think the user should also be told about this in the manual. Maybe include a limitations section in the manual? Again, it depends on what the user is using the code for. If it's some perovskite device with not many traps then this may not be an issue, but if it's some ultra trappy organic thing then.. ANSWER Yes, all charges are taken into account in computing the field (see eq. 2.1 in reference manual). C(x) denotes any other charges that may be present, such as doping, ions, trapped charge carriers. Would be nice if you could add some makefiles. Also would be good if you could tell me (the user) which package I need to install on Ubuntu to get pascal. I'm not super excited about installing free pascal binaries from anything but the Ubuntu repos due to security reasons. ANSWER We have outlined where to get fpc in the readme file. The compilation instructions are also simpler now and are reduced to a single statement, making it equally simple as a make file. On Wed, 20 Oct 2021 at 09:23, Marten Koopmans ***@***.***> wrote: > @UTH-Tuan <https://github.com/UTH-Tuan>, thank you for the info. It > seems that your Pascal compiler is not correctly installed. This does not > seem to be a problem with our codebase. I think that even a simple "Hello > world!" would not compile (as long as it requires system). We have had > students work with SIMsalabim on macOS before, and it compiled without > problems, but we don't have any macs around to debug the compiler > installation. > > You could try two things: > > - reinstall fpc (the compiler) > - install on Windows or Linux (it runs fine in a low-power VM, as the > program requires very little resources). Using Ubuntu (20.10 or newer) the > installation of the compiler is a simple one-liner, so that way you would > be sure that nothing can go wrong. For ubuntu, you would use sudo apt > install fpc to install the free pascal compiler. > > — > You are receiving this because you were mentioned. > Reply to this email directly, view it on GitHub > <[#3727 (comment)](https://github.com/openjournals/joss-reviews/issues/3727#issuecomment-947398235)>, > or unsubscribe > <https://github.com/notifications/unsubscribe-auth/AJFMMVH5TSAZVULWZBKMQYLUHZU5JANCNFSM5EAXX2FQ> > . > Triage notifications on the go with GitHub Mobile for iOS > <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675> > or Android > <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>. > >
