### 2021-11-15 #3727 @Mkoopm
@rkurchin: We are still working on some of the raised concerns.

@roderickmackenzie: We share your concerns about non-equilibrium traps and have made the bulk traps fully transient. We are working on the implementation of non-equilibrium interface traps and will send you an update when the changes are uploaded to GitHub.

@shahmoradi: We agree that testing is currently difficult. We are working on a python code that generates the graphs for all tests and also gives a pass/fail judgement. It should not take too long.
