+++ Paper/paper.bib
@@ -130,3 +130,65 @@ publisher = "AMER CHEMICAL SOC",
 number = "9",
 
 }
+
+﻿@Article{Courtier2019,
+author={Courtier, N. E.
+and Cave, J. M.
+and Walker, A. B.
+and Richardson, G.
+and Foster, J. M.},
+title={IonMonger: a free and fast planar perovskite solar cell simulator with coupled ion vacancy and charge carrier dynamics},
+journal={Journal of Computational Electronics},
+year={2019},
+month={Dec},
+day={01},
+volume={18},
+number={4},
+pages={1435-1449},
+abstract={Details of an open-source planar perovskite solar cell simulator, which includes ion vacancy migration within the perovskite layer coupled to charge carrier transport throughout the perovskite and adjoining transport layers in one dimension, are presented. The model equations are discretised in space using a finite element scheme, and temporal integration of the resulting system of differential algebraic equations is carried out in MATLAB. The user is free to modify device parameters, as well as the incident illumination and applied voltage. Time-varying voltage and/or illumination protocols can be specified, e.g. to simulate current--voltage sweeps, or to track the open-circuit conditions as the illumination is varied. Typical simulations, e.g. current--voltage sweeps, only require computation times of seconds to minutes on a modern personal computer. An example set of hysteretic current--voltage curves is presented.},
+issn={1572-8137},
+doi={10.1007/s10825-019-01396-2},
+url={https://doi.org/10.1007/s10825-019-01396-2}
+}
+
+
+@article{BURGELMAN2000,
+title = {Modelling polycrystalline semiconductor solar cells},
+journal = {Thin Solid Films},
+volume = {361-362},
+pages = {527-532},
+year = {2000},
+issn = {0040-6090},
+doi = {https://doi.org/10.1016/S0040-6090(99)00825-1},
+url = {https://www.sciencedirect.com/science/article/pii/S0040609099008251},
+author = {M Burgelman and P Nollet and S Degrave},
+keywords = {Solar cells, Thin film, Modelling, CdTe, Chalcopyrite},
+abstract = {An overview is given of various electronic effects present in polycrystalline thin film solar cells, which do not occur in standard crystalline Si solar cells. It is explained how these effects are treated numerically in a numerical solar cell simulation tool, SCAPS, developed at the author's institute. The capabilities and limitations of SCAPS are discussed. Simulation examples of current–voltage, capacitance–voltage and capacitance–frequency characteristics are given. The agreement between the simulations and measurements is shown and discussed, both for CdTe and for Cu(In,Ga)Se2 solar cells.}
+}
+
+
+@online{SETFOS,
+  %author = {MultiMedia LLC},
+  %title = {{MS Windows NT} Kernel Description},
+  %year = 2021,
+  url = {https://www.fluxim.com/setfos-intro},
+  urldate = {2021-11-19}
+}
+
+
+@online{gpvdm,
+  %author = {MultiMedia LLC},
+  %title = {{MS Windows NT} Kernel Description},
+  %year = 2021,
+  url = {https://www.gpvdm.com/},
+  urldate = {2021-11-19}
+}
+
+
+@online{driftfusion,
+  author = {P. Calado and I. Gelmetti and M. Azzouzi and B. Hilton and P.R.F. Barnes}
+  year = 2018,
+  url = {https://github.com/barnesgroupICL/Driftfusion},
+  urldate = {2021-11-19}
+}
+
