### 2021-11-23 9792dcf
remove old SIMsalabim folder
