### 2021-11-23 #3727 @ljakoster
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Mon, 15 Nov 2021 at 14:15, Marten Koopmans ***@***.***>
wrote:

> @rkurchin <https://github.com/rkurchin>: We are still working on some of
> the raised concerns.
>
> @roderickmackenzie <https://github.com/roderickmackenzie>: We share your
> concerns about non-equilibrium traps and have made the bulk traps fully
> transient. We are working on the implementation of non-equilibrium
> interface traps and will send you an update when the changes are uploaded
> to GitHub.
>
> @shahmoradi <https://github.com/shahmoradi>: We agree that testing is
> currently difficult. We are working on a python code that generates the
> graphs for all tests and also gives a pass/fail judgement. It should not
> take too long.
>
