+++ Paper/paper.bib
@@ -168,18 +168,12 @@ abstract = {An overview is given of various electronic effects present in polycr
 
 
 @online{SETFOS,
-  %author = {MultiMedia LLC},
-  %title = {{MS Windows NT} Kernel Description},
-  %year = 2021,
   url = {https://www.fluxim.com/setfos-intro},
   urldate = {2021-11-19}
 }
 
 
 @online{gpvdm,
-  %author = {MultiMedia LLC},
-  %title = {{MS Windows NT} Kernel Description},
-  %year = 2021,
   url = {https://www.gpvdm.com/},
   urldate = {2021-11-19}
 }
