+++ Paper/paper.bib
@@ -180,7 +180,7 @@ abstract = {An overview is given of various electronic effects present in polycr
 
 
 @online{driftfusion,
-  author = {P. Calado and I. Gelmetti and M. Azzouzi and B. Hilton and P.R.F. Barnes}
+  author = {P. Calado and I. Gelmetti and M. Azzouzi and B. Hilton and P.R.F. Barnes},
   year = 2018,
   url = {https://github.com/barnesgroupICL/Driftfusion},
   urldate = {2021-11-19}
