### 2021-11-23 #3727 @kostergroup
@roderickmackenzie
Your use of the V in equations 2.1 -2.6 is a bit confusing because commonly V is only used in the DD equations when hetero-junctions are not taken into account. When the model has been adjusted to handle hetero-junctions (as yours seems to be 2.13-2.14) it's conmen to write Ec and Ev in the DD equations instead of V, to explicitly say that heterosteps can drive current. This is especially confusing as your fig 2.1 shows hetero steps so made me scratch me head as soon as I started reading until I reached the interfaces section. If you rewrite in terms of Ec and Ev you can get rid of equations 2.13-2.14, it would make the description look more like the classical DD literature as used for things like the III-V materials - and make things a bit clearer.

ANS:
Fair point. We have chosen a slightly different way of presenting the DD equations to 1) improve the connection to other literature, while 2) remaining close to what is in the actual code: In the code, one simply replace V(x) in the DD equations by generalized potentials (see K. Zojer et al., Organic Electronics 11 (2010) 1999–2011 and the book by Snowden (ref in reference manual)) that include both the offsets in Ec/Ev as well as any changes in the effective DOS. Both points 1 and 2 are relevant here. So, we have modified the description of the DD equations (in the reference manual) to stress that these take the generalized potentials (Vgn, Vgp) in case there are offsets in Ec/v or the effective DOS. These reduce to V if there are no heterojunctions/internal interface, and refer to section 2.3 to explain how Vgn and Vgp are obtained if Ec,Ev,Ncv are not constant.

As I understand it there are two parts to the model a transient part and a steady state part. In the intro you talk about transient photovoltage and AC analysis. However your equation 2.22 is the steady state form of the SRH equation, so you are assuming your trapped carrier distribution is in quasi-steady state. This means that although you can do time domain simulations your trapped carriers will always think they are in quasi-equilibrium. This won't matter for DC or for slow small perturbation stuff. But you won't for example be able to simulate things far from equilibrium, e.g. ToF transients are an extreme example of this where carriers are initially captured then gradually released from deeper and deeper traps as time goes on. TPV transients could also meet this large signal condition depending on the power of the laser pulse, and AC depending on carrier trapping/detrapping rates and the frequency you are at (and trap depth). CELIV will also fall into large perturbation regime etc... etc... I think the user should be alerted to the fact that your carriers are in SS and the model does not explicitly solve the SRH equations in time domain. Many users won't grasp this so I think it needs to be clearly pointed out and what the implications of this are for results.
I think my main concern with points 2 and 3 is that the user knows what he/she is dealing with. There are a tonne of papers where people just use software (aka COMSOL etc..) to produce a paper without understand what is going on under the hood and results in miss interpretation of data which is not a good thing.

ANS
We fully agree, and this is one of the main reasons for making this project open-source and for spending the time to write a reference manual. Ultimately, we decided that it was less work to make it fully transient instead of trying to precisely define when our original approach worked/did not work (especially since this had been on the wishlist). We have added a description of how the trapping and detrapping is accounted for (reference manual).

QUESTION:
We would like to add a reference to your code (in the paper), how would you like to be cited?

@UTH-Tuan
1) We have included references to commercial, freeware, and open-source alternatives to SIMsalabim, as requested.
2) In order to better highlight the relation between the SIMsalabim project as a whole and the steady-state and transient flavours, we have renamed the steady-state code (SimSS, simulates steady-state).
3) We have made the tests fully automated!
