### 2021-12-02 #3727 @roderickmackenzie
@kostergroup
 Thanks for your comments and work. And apologies for the delay in getting back to you I was taken out by covid for a while.  Comments:

*So in the manual I think the section on trapping should be separated from the potential section (2.1.1) and given it’s own heading.   I think in the trapping section you should explicitly call it SRH trapping/recombination – at the moment you cite the SR paper but don’t call it SRH. And possibly move the steady state SRH equation up into this section and explain the relationship between these two approaches. At the moment the casual reader may not make the connection.

*I think you need to write how Cn/Cp are calculated from SRH theory.

*You asked “We would like to add a reference to your code (in the paper), how would you like to be cited?” I presume this was aimed at me – thanks :). I think it would be good if you could cite https://doi.org/10.1002/aenm.201100709 in the trapping section of the manual as doing dynamic SRH trapping for organics in this way was first demonstrated there. Maybe also cite the webpage www.gpvdm.com if you want to cite the code base.

*Now, I don’t want to be overly critical and I appreciate the effort you went to, to put in traps. And a model with traps is infinity better than a model with no traps. But I still don’t think you can simulate a TPC transient. But, I would be interested to see if you can reverse bias your device to say -2V then apply a laser pulse and then from the transient extract say a 100meV Urbach energy as done here https://doi.org/10.1021/jp4010828. This is the acid test of a transient model.

The reason I don’t think you can do this is because my reading of the manual is that you have a single trap level with one capture/escape cross section. In reality the *escape* coefficients will change depending on the energetic depth of the trap. (See eqn. 2.9 in the original SR paper) This is the reason TPC transients look as they do.  Early current comes from shallow traps deep current comes from deep traps.  Also my reading of your manual is that you have a single trap density Ntb, rather than a distribution of traps over energy space which is another reason you won’t be able to model TPC transients.  Furthermore, you are assuming that all traps will have the same occupational probability ftb which is true as you only have one trap but in reality there is no reason why shallow traps should have the same occupation as deep traps.

Furthermore, for a given trap energy range dE your occupation probability ftb, will not be a single number but a range of numbers dictated by fermi-dirrac statistics.  So all trap information such as density of carriers/escape rates etc.. have to be calculated by integrating the fermi-dirrac occupation function multipled by a density of states which also changes over energy space. (say Nexp(-E/Eu)). Thus you should really be solving for Fermi-level rather than ftb directly. I suspect you need to do this last step to get smooth TPC transients.

If this were an APL letter, I would probably not bother pushing this point. However, as I mentioned before my concern with publishing models, is they tend to be used by other people who don’t understand the bounds of validity of the model. So either I would I would like you to simulate a TPC transient and extract the Urbach energy out of it and demonstrate this in the manual as a test case.  Or very clearly say where you think your model is valid and where it is not.

For example, I’m fairly sure it won’t be able to simulate ToF/TPC transients – if my reading of the manual is correct.  I also suspect as it is, it won’t be able to simulate CELIV transients very well as multiple trap levels are required for those simulations (https://doi.org/10.1063/1.4818267), I would also have doubts about simulating TPV transients with such a model because there is no reason why in a fast experiment such as TPV deep and shallow traps should have a single occupation probability.  I do think the model is valid for slow scans of perovskite JV curves though.
