+++ Paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 The drift-diffusion technique is well-established way of describing charge transport in semiconductors. While in some special cases it is possible to analytically solve the problem, typically numerical solutions are the best achievable result, especially in more complicated semiconductor devices, such as solar cells.
 Numerical solutions can be found however and often prove to closely align with experimental observations.
 
-SIMsalabim is a  software that allows researchers to simulate solar cells or other semiconductor devices to access parameters that might be hard to observe in an experiment. SIMsalabim can do both steady-state and time-dependent simulations including effects of illumination, mobile ions, recombination, trapping, and dielectric mismatch. While SIMsalabim is not restricted to the simulation of solar cells, some relevant solar cell experiments that can be simulated include: transient photovoltage, current-voltage sweeps (including hysteresis).
+SIMsalabim is a  software that allows researchers to simulate solar cells or other semiconductor devices to access parameters that might be hard to observe in an experiment. SIMsalabim can do both steady-state and time-dependent simulations including effects of illumination, mobile ions, recombination, trapping, and dielectric mismatch. While SIMsalabim is not restricted to the simulation of solar cells, some relevant solar cell experiments that can be simulated include: transient photovoltage, current-voltage sweeps (including hysteresis). SIMsalabim can be used to simulate many different device types, but the user should make sure that the physical model implemented in SIMsalabim fits the device and operating regime that is simulated. Because the physical model is constantly evolving we refer the reader to the manual for what is currently implemented.
 
 # Statement of need
 SIMsalabim is a numerical 1D drift-diffusion simulation software implemented in Pascal. Pascal enables a combination of speed offered by low-level languages (e.g. C) while enforcing code readability and simplicity. The code is therefore relatively accessible to the typical material scientists who only have experience in higher-level languages (e.g. Python, MATLAB). SIMsalabim allows setting of parameters either in a file or the command line, meaning that it is both scriptable and user-friendly. The scriptable nature allows for automatised fitting of experiments, or scanning of extremely large parameter spaces [@LeCorre2021Feb].
