+++ Paper/paper.bib
@@ -168,12 +168,14 @@ abstract = {An overview is given of various electronic effects present in polycr
 
 
 @online{SETFOS,
+	author = {FLUXiM},
   url = {https://www.fluxim.com/setfos-intro},
   urldate = {2021-11-19}
 }
 
 
 @online{gpvdm,
+  author = {R. MacKenzie},
   url = {https://www.gpvdm.com/},
   urldate = {2021-11-19}
 }
