### 2021-12-09 #3727 @kostergroup
*So in the manual I think the section on trapping should be separated from the potential section (2.1.1) and given its own heading. I think in the trapping section you should explicitly call it SRH trapping/recombination – at the moment you cite the SR paper but don’t call it SRH. And possibly move the steady state SRH equation up into this section and explain the relationship between these two approaches. At the moment the casual reader may not make the connection.

**ANS**
We, as per your request, specified explicitly which terms enter C(x) (in the Poisson equation, eq 2.1) and how this term (C(x)) is computed. Indeed, this part of the manual was not very clear. So, we have renamed section 2.5 ‘Recombination and Trapping’. In this section, we specify how the occupancy of the trap level and the capture/emission rates are computed in S-S and in the transient case.

*I think you need to write how Cn/Cp are calculated from SRH theory.

**ANS**
Our Cn and Cp are simply the capture coefficients as related to the capture cross section (sigma) and the carrier velocity. So Cn/p are not calculated, but are input parameters and their unit is volume/time, as used in for example doi: 10.1038/srep37511 or in textbooks like eqs (7-155) and (7-156) in K.C. Kao, Dielectric phenomena in Solids.
*Now, I don’t want to be overly critical and I appreciate the effort you went to, to put in traps. And a model with traps is infinitely better than a model with no traps. But I still don’t think you can simulate a TPC transient. But, I would be interested to see if you can reverse bias your device to say -2V then apply a laser pulse and then from the transient extract say a 100meV Urbach energy as done here https://doi.org/10.1021/jp4010828. This is the acid test of a transient model.

**ANS:**
I think there is some misunderstanding about what we did in response to your previous comments: Trapping was included from the start. However, as you pointed out correctly, the transient code (ZimT) did use a quasi-static approach to trapping rather than a fully transient one. So, we changed the way we calculate the trap occupancy. Instead of relying on an (artificial) equilibrium between trapped and free charges (quasi-S-S), we now use capture and emission rates to explicitly solve for the trap occupancy.
We did not, however, introduce elaborate distributions of trapping levels in energy.

The reason I don’t think you can do this is because my reading of the manual is that you have a single trap level with one capture/escape cross section. In reality the escape coefficients will change depending on the energetic depth of the trap. (See eqn. 2.9 in the original SR paper) This is the reason TPC transients look as they do. Early current comes from shallow traps deep current comes from deep traps. Also my reading of your manual is that you have a single trap density Ntb, rather than a distribution of traps over energy space which is another reason you won’t be able to model TPC transients. Furthermore, you are assuming that all traps will have the same occupational probability ftb which is true as you only have one trap but in reality there is no reason why shallow traps should have the same occupation as deep traps.

**ANS:**
No, we are not assuming that ftb is static. We solve for ftb explicitly for every step and grid point. (Just as an <aside>: ftb is indeed not a constant and this becomes obvious by simulating TPC with a single trap level and enough traps to cause significant build-up of space charge. The resulting TPC transient is pretty much like a power-law and thus shows a multitude of lifetimes despite the single trap level:
![tpc](https://user-images.githubusercontent.com/64084163/145418974-284f1d99-70fb-4b2a-bb0f-ac9f8292c858.png)

<\aside>)
Furthermore, for a given trap energy range dE your occupation probability ftb, will not be a single number but a range of numbers dictated by fermi-dirrac statistics. So all trap information such as density of carriers/escape rates etc.. have to be calculated by integrating the fermi-dirrac occupation function multipled by a density of states which also changes over energy space. (say Nexp(-E/Eu)). Thus you should really be solving for Fermi-level rather than ftb directly. I suspect you need to do this last step to get smooth TPC transients.

**ANS**
f_tb is indeed obtained from F-D statistics (in steady-state of course). However, we only have a single trap level as pointed out in the manual. Extending the code to more elaborate trap distributions (in energy and / or space) is not difficult, yet beyond the scope of the present work.

If this were an APL letter, I would probably not bother pushing this point. However, as I mentioned before my concern with publishing models, is they tend to be used by other people who don’t understand the bounds of validity of the model. So either I would I would like you to simulate a TPC transient and extract the Urbach energy out of it and demonstrate this in the manual as a test case. Or very clearly say where you think your model is valid and where it is not.

**ANS**
Yes, we agree. This is crucial, as we have stated before, and one of the motivations for making this open-source rather than freeware. The assumptions underlying the codes are outlined in the part of the manual that lists the basic equations. However, some users might not be so keen on looking at these equations so we have summarized what is in the model in the beginning of chapter 2 (“What is currently included?”). Here, among other things, we explicitly state that we have one 1 trap level.
The responsibility for assessing whether SIMsalabim is suited to describe experiments on some materials system lies squarely with the user (we also point this out in the manual and paper). We are not claiming to have the perfect physical model (drift-diffusion has plenty of assumptions, no matter how sophisticated). Rather, we are trying to get a software package out there including a description of the underlying assumptions. It is up to the users (if any:)) to verify whether the model is applicable or not. If the worst comes to the worst and a user tries to describe the experiments you’re referring to, they will find that the model does not fit the data.

We would like to point out that we are not advocating a model. We are publishing a software package that is capable of simulating the model described in the manual, but this model is constantly evolving and has been for a long time. Publishing physical models is not the scope of the Journal of Open Source Software. We would therefore like to keep the discussion focused on whether the code, documentation, and testing is of sufficient quality. We think our job here is to clearly explain our model in the documentation.

All modelling software is incomplete in terms of the physics and one can always find systems and experiments that cannot be explained/fitted. However, that does not mean that the software itself is meaningless. As we understand it, JOSS publishes articles to guarantee that the software does what it is supposed to do, there is adequate documentation and testing. Our software, unless we are missing something, does indeed function as specified in the manual, it is documented, and we have automated tests in place. Previously (when using the quasi-static trapping/detrapping), the software did indeed not fully function as suggested in the manual. Now, however, this is no longer the case.
