### 2021-12-16 #3727 @rkurchin
@UTH-Tuan: what's the status of your last couple checklist items?

@Mkoopm: how about the updated testing in response to @shahmoradi's concern? (I see some changes have been made in the tests directory, but at least `run_tests.py` appears not to have been updated in several weeks)

@roderickmackenzie: What are your thoughts on @ljakoster's last message?

To add a bit of editorial perspective on this last aspect, I agree with Prof. Koster that JOSS's primary job is to review for whether the software does what it says it does, but I also agree with Prof. Mackenzie's perspective that any tool should be "clearly labeled" as to its appropriate uses. So, acknowledging the age-old Box quote, hopefully we can come to some consensus about where this model is useful, where it might be particularly wrong, and how it ought to be labeled accordingly. Thanks again to all for your detailed, thoughtful, and good-faith engagement in this process! 😄
