### 2021-12-16 #3727 @Mkoopm
@rkurchin: the script `run_tests.py` was added responding to @shahmoradi's concerns, it automates the testing procedure and not been updated because it works (tests can now be automatically run).
