### 2021-12-17 #3727 @shahmoradi
The test script `run_tests.py` works on Windows OS. It failed on MingW and Linux OS with various library errors. I assume these errors are due to version incompatibilities. While I am not going to require the authors to make this extra change in the tests, I highly recommend adding a few lines at the beginning of the script to automatically check proper library versions are installed on the user's system before running the tests. This way, it will be clear what exactly the source of test failures is, rather than getting highly obscure errors from random dependency libraries in the Python session.

I have check-marked the last remaining item in **Review checklist for @shahmoradi**.
