### 2021-12-20 #3727 @Mkoopm
@shahmoradi, thank you for the thorough checking. We will include a nicer error message with an installation suggestion in the next update. Did you follow the instructions in `tests.md`, or did you just try to run `run_tests.py`? If the instructions don't work on your systems I will look into the installation method as well.
