### 2021-12-20 #3727 @roderickmackenzie
In my last review I asked you to “...to see if you can reverse bias your device to say -2V then apply a laser pulse and then from the transient extract say a 100meV Urbach energy as done here https://doi.org/10.1021/jp4010828. This is the acid test of a transient model.” But you have not done this, so have done it for you using the graph you provided above. The results are below.  The graph plots the normalized extracted density of states against energetic depth going into the band (0eV = close to band edge, 0.2 eV = deep into the band). I have included what one would expect for Urbach energies of 100 meV, 60 meV and 30 meV. One would expect a DoS descending into the bandgap. Also plotted in the data set is the DoS extracted from your transient using equation 1 and 2 in the paper mentioned above. You will notice that the line generated from your data has the wrong gradient, such a graph would indicate the trap density increases going into the bad which is clearly wrong.

![image](https://user-images.githubusercontent.com/2600516/146774757-bc4002d8-cb84-4488-84a8-d3546adf2ff0.png)

The reason the extracted DoS goes the wrong way is because your model only has one trap state and you don’t discretized your Urbach tail in energy space.  Urbach energy is the *defining feature* in many time domain transients, TPC, ToF, TPV and many others. See for example the discussions in (Seynhaeve and  Main wrote a lot on this topic):

*C. Main, D. Nesheva, Transient photocurrent techniques as a mean of characterising amorphous semiconductors
*Seynhaeve, Post-transit time-of-Aight currents as a probe of the density of states in hydrogenated amorphous silicon
*C. Main, Interpretation of photocurrent transients in amorphous semiconductors Street, Localized state distribution and its effect on recombination in organic solar cells
*C. Main, Computer Modellin gof Multi-Trapping and hopping transport in disordered semiconductors.

You say in the paper “While SIMsalabim is not restricted to the simulation of solar cells, some relevant solar cell experiments that can be simulated include: transient photovoltage, current-voltage sweeps (including hysteresis). SIMsalabim can be used to simulate many different device types, but the user should make sure that the physical model implemented in SIMsalabim fits the device and operating regime that is simulated.”

I find this problermatic (especially for someone inexperienced), because on one hand you say it’s time domain but on the other you are saying be careful as the model may not be correct. In reality people are going to use this model for disordered solar cells, such as in OPV or perovskite because that’s what the community is working on. In these systems you you need the Urbach tail discritized in energy space or you will get results as shown above.

This for me this is the fundamental sticking point. It claims to be time domain model but it’s not clear which time domain experiments it’s valid for. I would not want to be the PhD student who tries to use this model to simulate anything in time domain. It could at best result in frustration and at worst result in papers in the literature which should not be published. So in terms of JOSS criteria I can’t tick the box “Has a clear description of the high-level functionality and purpose of the software for a diverse, **non-specialist** audience been provided”

From my point of view there are two options:

1. Write a paragraph in the manual properly reflecting about where/when the model can be used and where it can not be used. For example I would not use it to simulate the post-transit regime in TPC or ToF transients. I would worry about TPV transients especially anything which was non mono-exponentiation. CELIV may be OK. Slow Perovskite JV scans would prob be OK although fast scans may not be. I think you need to reflect on this an be up front about the user about limitations esp. for an "**non-specialist** audience". This is not a big ask and should take about half an hour to write.

2. If you are going to claim it’s time domain in the way you do without caveats: Implement Urbach tails properly and demonstrate you can simulate TPC transients then extract the Urbach energy using the methods presented  https://doi.org/10.1021/jp4010828, and developed by Main et al.

This reviewer will simply not let the paper pass without condition 1 or 2 being met. I don’t think 1 is a big ask...
