### 2021-12-21 #3727 @kostergroup
@roderickmackenzie
In my last review I asked you to “...to see if you can reverse bias your device to say -2V then apply a laser pulse and then from the transient extract say a 100meV Urbach energy as done here https://doi.org/10.1021/jp4010828. This is the acid test of a transient model.” But you have not done this, so have done it for you using the graph you provided above.

ANS
We did not do this as this would obviously fail.

From my point of view there are two options:
    1. Write a paragraph in the manual properly reflecting about where/when the model can be used and where it can not be used. For example I would not use it to simulate the post-transit regime in TPC or ToF transients. I would worry about TPV transients especially anything which was non mono-exponentiation. CELIV may be OK. Slow Perovskite JV scans would prob be OK although fast scans may not be. I think you need to reflect on this an be up front about the user about limitations esp. for an "non-specialist audience". This is not a big ask and should take about half an hour to write.
    2. If you are going to claim it’s time domain in the way you do without caveats: Implement Urbach tails properly and demonstrate you can simulate TPC transients then extract the Urbach energy using the methods presented https://doi.org/10.1021/jp4010828, and developed by Main et al.

This reviewer will simply not let the paper pass without condition 1 or 2 being met. I don’t think 1 is a big ask...

ANS
Both options require quite a bit of thought, not in the least as I anticipate that we would not agree straight away about option 1. So, please bear with us for a bit; we will get back to you in the new year.
