### 2022-01-03 #3727 @rkurchin
Happy New Year everyone! @kostergroup, just a friendly reminder to update on this when you've decided your next course of action. I recognize some folks are still on holiday this week, and that as you said, it may require substantial thought. Depending on how long you want, I can also "pause" this review for a few weeks if you like.
