### 2022-01-10 #3727 @kostergroup
Happy 2022 everyone!

@rkurchin I've just returned to the office and I think we'll get back to
this in a week or 2. Not sure that qualifies as a 'pause'? probably.

jan anton

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Mon, 3 Jan 2022 at 22:52, Rachel Kurchin ***@***.***>
wrote:

> Happy New Year everyone! @kostergroup <https://github.com/kostergroup>,
> just a friendly reminder to update on this when you've decided your next
> course of action. I recognize some folks are still on holiday this week,
> and that as you said, it may require substantial thought. Depending on how
> long you want, I can also "pause" this review for a few weeks if you like.
>
