### 2022-01-24 #3727 @ljakoster
@rkurchin: thanks. We're almost done!

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Sun, Jan 23, 2022 at 7:54 PM Rachel Kurchin ***@***.***>
wrote:

> Hi again @kostergroup <https://github.com/kostergroup> @ljakoster
> <https://github.com/ljakoster>, just checking in on this!
>
