### 2022-01-25 #3727 @kostergroup
@roderickmackenzie

> From my point of view there are two options:
> 1 Write  …  write.
> 2
> If you are going to claim it’s time domain in the way you do without caveats: Implement Urbach tails properly and demonstrate you can simulate TPC transients then extract the Urbach energy using the methods presented https://doi.org/10.1021/jp4010828, and developed by Main et al.
>
> This reviewer will simply not let the paper pass without condition 1 or 2 being met.

ANS

OK, thanks for bearing with us. We have opted for option 2. So, we have implemented the possibility of defining multiple trap levels (either bulk or interfacial). To validate the approach, we have put in a distribution of bulk trap levels below the conduction band that follows an Urbach tail of 100 meV and analysed the data as per your publication. We get the following:

So below the conduction band we indeed find a very nice exponential tail with an Urbach energy of 95 meV, very close to the input. (this is test 11)

![test_11_result](https://user-images.githubusercontent.com/64084163/150976715-d11e9b0d-0b3a-406c-8fe2-c7ceec5a0812.png)

For the bulk traps, we use:
![papers_per_year1](https://user-images.githubusercontent.com/64084163/150976822-b182964f-fcf9-43f8-a33f-4a6c79a1dcff.png)

Where the conduction band sits at 3.0 eV.
We trust that we have met your request.

@shahmoradi: We added a check for python dependencies to the testing code. Now, installation instructions are shown if a dependency is not met.
