### 2022-01-25 #3727 @themagicrobot
Good effort! And thanks for your efforts. I'm happy for this to be published. My only final suggestion is that it would be good if you could put a paragraph in the manual saying that a distribution of traps is important for disordered materials. So people don't get tempted to turn it off. best, Rod
