### 2022-01-25 #3727 @roderickmackenzie
Good effort! And thanks for all your work. My only final suggestion is that you put a paragraph in the manual saying that a distribution of trap states are important when simulating disordered materials. That way people won't be tempted to run simulations/publish papers with no traps. best, Rod
