### 2022-01-26 #3727 @kostergroup
@roderickmackenzie
Good effort! And thanks for all your work. My only final suggestion is that you put a paragraph in the manual saying that a distribution of trap states are important when simulating disordered materials. That way people won't be tempted to run simulations/publish papers with no traps. best, Rod

ANS

Thanks for your final suggestion; this is our final modification. You provided us with two options; we opted for the more laborious option 2 as it extends the code, which is time well spent. By asking us to specify when distributions of traps are important, you are effectively asking us to also implement option 1. In all fairness, this is asking too much.

However, we have changed the manual to warn the user of the importance of getting the trapping right. Firstly, we have renamed section ‘Recombination and Trapping’ => ‘Trapping and Recombination’ to put the emphasis on trapping. Next, we have added to the manual (section ‘Trapping and Recombination’):

> By their very nature, traps capture charge carriers. It is critical to understand their effects on a device in order to judge their relevance. Trapped charge carriers contribute to the space charge (see Eq. 2.1), without contributing to charge transport. Additionally, trap levels introduce another recombination pathway. Both effects are typically bad for device performance.
>
> In transient simulations, things get a little bit more complicated as the trapping and/or detrapping can take rather long and, therefore, it might take a long time (as compared to the simulated time) before equilibrium is reached. This is especially relevant to cases where there is a distribution of traps, giving rise to multiple trapping/detrapping times.
>
> In sum, traps can have a large impact on the simulations, depending on what is simulated and in what regime.
>
