### 2022-02-04 #3727 @ljakoster
@rkurchin We've received a proof from whedon; does that mean the manuscript
has been accepted for publication?

best wishes,
Jan Anton
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>
