### 2022-02-04 #3727 @rkurchin
Jan Anton,

I generated the latest proof so I can do the last few editorial checks while I wait to hear from the other two reviewers that they're happy to accept. The other things you'll need to do before we finalize this submission are (assuming no other changes are needed):

1. Merge any and all changes from this review into your main branch and **issue a new version tag**. _(If you want to merge in the paper, you may, but it is not required that the actual manuscript live into the repo in perpetuity since JOSS will host it and you can simply add a badge link or whatever you like. But the actual changes to software and docs do need to be merged!)_
2. Create a DOI for the contents of the repo at the same commit corresponding to that version tag, e.g. using figshare or Zenodo. Please make sure that the metadata (version number, title, author list, etc.) match those of your manuscript.
3. Post a comment here with the version number and DOI.
