### 2022-02-04 #3727 @rkurchin
Here are some small editorial comments on the manuscript itself:
* line 10: there should be commas around "however" i.e. "numerical solutions can be found, however, and..."
* 17: since this list is only two items, I would remove the colon and just phrase it as: "...include transient photovoltage and current-voltage sweeps (including hysteresis)."
* 20: add comma after "evolving"
* 28: "automatised" --> "automated"
* 32-34: I'm not entirely certain what this sentence is saying. Can you rephrase/clarify?
* 39-40: add "of" between "combination" and "features"
* 45-47: add spaces between software names and citations
