### 2022-02-04 #3727 @rkurchin
@ljakoster please also fix that one invalid DOI ;)
