### 2022-02-08 #3727 @ljakoster
Dear All,

great! I would like to thank the reviewers for their patience and input;
the project has greatly improved as a result of it.

@rkurchin: Marten and I are working on this.

best wishes,
Jan Anton
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Mon, Feb 7, 2022 at 10:26 PM Rachel Kurchin ***@***.***>
wrote:

> Great, so @ljakoster <https://github.com/ljakoster>, once the couple last
> things above are addressed, this is ready to rock! 🚀
>
