+++ Paper/paper.bib
@@ -159,7 +159,7 @@ volume = {361-362},
 pages = {527-532},
 year = {2000},
 issn = {0040-6090},
-doi = {https://doi.org/10.1016/S0040-6090(99)00825-1},
+doi = {10.1016/S0040-6090(99)00825-1},
 url = {https://www.sciencedirect.com/science/article/pii/S0040609099008251},
 author = {M Burgelman and P Nollet and S Degrave},
 keywords = {Solar cells, Thin film, Modelling, CdTe, Chalcopyrite},
