### 2022-02-08 66198b5
Remove paper source files from master branch
