+++ Paper/paper.bib
@@ -1,190 +1,0 @@
-@article{LeCorre2021Feb,
-	author = {Le Corre, Vincent M. and Sherkar, Tejas S. and Koopmans, Marten and Koster, L. Jan Anton},
-	title = {{Identification of the dominant recombination process for perovskite solar cells based on machine learning}},
-	journal = {Cell Rep. Phys. Sci.},
-	volume = {2},
-	number = {2},
-	year = {2021},
-	month = {Feb},
-	issn = {2666-3864},
-	publisher = {Elsevier},
-	doi = {10.1016/j.xcrp.2021.100346}
-}
-
-@article{SpaceCharge2021,
-author = {Le Corre, Vincent M. and Duijnstee, Elisabeth A. and El Tambouli, Omar and Ball, James M. and Snaith, Henry J. and Lim, Jongchul and Koster, L. Jan Anton},
-title = {Revealing Charge Carrier Mobility and Defect Densities in Metal Halide Perovskites via Space-Charge-Limited Current Measurements},
-journal = {ACS Energy Letters},
-volume = {6},
-number = {3},
-pages = {1087-1094},
-year = {2021},
-doi = {10.1021/acsenergylett.0c02599},
-URL = {https://doi.org/10.1021/acsenergylett.0c02599},
-eprint = {https://doi.org/10.1021/acsenergylett.0c02599}
-}
-
-@Article{Friso2020,
-author ="Hu, Dingqin and Yang, Qianguang and Chen, Haiyan and Wobben, Friso and Le Corre, Vincent M. and Singh, Ranbir and Liu, Tao and Ma, Ruijie and Tang, Hua and Koster, L. Jan Anton and Duan, Tainan and Yan, He and Kan, Zhipeng and Xiao, Zeyun and Lu, Shirong",
-title  ="15.34% efficiency all-small-molecule organic solar cells with an improved fill factor enabled by a fullerene additive",
-journal  ="Energy Environ. Sci.",
-year  ="2020",
-volume  ="13",
-issue  ="7",
-pages  ="2134-2141",
-publisher  ="The Royal Society of Chemistry",
-doi  ="10.1039/D0EE00714E",
-url  ="http://dx.doi.org/10.1039/D0EE00714E",
[…]
