### 2022-02-08 #3727 @ljakoster
@rkurchin:

I think we have done all the things you requested. We have a doi:
https://doi.org/10.5281/zenodo.6010442
All the changes already were in the latest version (4.29) and we have
created a new version tag: v4.29-JOSS
<https://github.com/kostergroup/SIMsalabim/releases/tag/v4.29-JOSS>.

Thanks for editing the paper; we've modified the text. Please let me know
if we've missed something.

best wishes,
Jan Anton

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Fri, Feb 4, 2022 at 12:42 PM Rachel Kurchin ***@***.***>
wrote:

> Jan Anton,
>
> I generated the latest proof so I can do the last few editorial checks
> while I wait to hear from the other two reviewers that they're happy to
> accept. The other things you'll need to do before we finalize this
> submission are (assuming no other changes are needed):
>
>    1. Merge any and all changes from this review into your main branch
>    and *issue a new version tag*. *(If you want to merge in the paper,
>    you may, but it is not required that the actual manuscript live into the
>    repo in perpetuity since JOSS will host it and you can simply add a badge
>    link or whatever you like. But the actual changes to software and docs do
>    need to be merged!)*
>    2. Create a DOI for the contents of the repo at the same commit
>    corresponding to that version tag, e.g. using figshare or Zenodo. Please
>    make sure that the metadata (version number, title, author list, etc.)
>    match those of your manuscript.
>    3. Post a comment here with the version number and DOI.
>
