### 2022-02-08 #3727 @rkurchin
Dear @ljakoster and reviewers @UTH-Tuan, @shahmoradi, and @roderickmackenzie,

Whew, this review has been a bit of a saga! Thank you all for your continued thoughtfulness, engagement, courtesy, and good faith throughout the process. I think everyone can agree the submission improved substantially (both in functionality and clarity of messaging), which is really the goal and gold standard of JOSS reviews. I'm so happy to have been able to edit and help shepherd this through, and congratulations!

Rachel
