### 2022-02-09 #3727 @ljakoster
@rkurchin: we have checked the proof and it looks perfect, thanks!
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>
