### 2022-02-09 #3727 @kthyng
Hi @ljakoster! A few last items:

1. zenodo and version look good.
2. I recommend putting `` around `SIMsalabim` so it stands out as software.
3. Is Hu et al missing a title?
4. MacKenzie et al seems like it is missing information?
5. Do your references just happen to stop at M? Seems odd.
