### 2022-02-10 #3727 @ljakoster
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Thu, Feb 10, 2022 at 12:37 AM Kristen Thyng ***@***.***>
wrote:

> Hi @ljakoster <https://github.com/ljakoster>! A few last items:
>
>    1. zenodo and version look good.
>    2. I recommend putting `` around SIMsalabim so it stands out as
>    software.
>    3. Is Hu et al missing a title?
>    4. MacKenzie et al seems like it is missing information?
>    5. Do your references just happen to stop at M? Seems odd.
>
