### 2022-02-10 #3727 @kostergroup
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Thu, Feb 10, 2022 at 10:33 AM Jan Anton Koster ***@***.***>
wrote:

> @whedon <https://github.com/whedon> generate pdf from branch paper
>
>
> --
> Prof. Dr. L.J.A. (Jan Anton) Koster
> Zernike Institute for Advanced Materials
> University of Groningen
> open-source simulator for perovskite and organic PV
> <https://github.com/kostergroup/SIMsalabim>
> group website <http://www.koster-group.nl>
>
>
> On Thu, Feb 10, 2022 at 12:37 AM Kristen Thyng ***@***.***>
> wrote:
>
> > Hi @ljakoster <https://github.com/ljakoster>! A few last items:
> >
> > 1. zenodo and version look good.
> > 2. I recommend putting `` around SIMsalabim so it stands out as
> > software.
> > 3. Is Hu et al missing a title?
> > 4. MacKenzie et al seems like it is missing information?
> > 5. Do your references just happen to stop at M? Seems odd.
> >
> > —
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/3727#issuecomment-1034315422
> >,
> > or unsubscribe
> > <
> https://github.com/notifications/unsubscribe-auth/AJFMMVE6EUFWB6V52NO7JZLU2L3BVANCNFSM5EAXX2FQ
> >
> > .
> > Triage notifications on the go with GitHub Mobile for iOS
> > <
> https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675
> >
> > or Android
> > <
> https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub
> >.
> >
> > You are receiving this because you were mentioned.Message ID:
> > ***@***.***>
> >
>
