### 2022-02-10 #3727 @ljakoster
Dear Kristen,

thanks for your email.

2: thanks, done.

3: yes, this was missing; we've added the title, looks like this was a
BibTeX issue.

4: This is a reference to an internet site. We've asked this author (= one
of the reviewers) how they would like to be cited and they have provided
this reference.

5: that is odd, yet correct!

best wishes,

Jan Anton
--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>
