### 2022-02-11 #3727 @kostergroup
Dear Kristen,

ah, sorry, we misunderstood. We've changed all mentions of SIMsalabim to
the `` version.

best wishes,
Jan Anton

--
Prof. Dr. L.J.A. (Jan Anton) Koster
Zernike Institute for Advanced Materials
University of Groningen
open-source simulator for perovskite and organic PV
<https://github.com/kostergroup/SIMsalabim>
group website <http://www.koster-group.nl>

On Thu, Feb 10, 2022 at 9:09 PM Kristen Thyng ***@***.***>
wrote:

> @ljakoster <https://github.com/ljakoster> ok! One thing — I meant ``
> around your software name throughout the paper, but it is up to you. Would
> you like to change that before we accept the final version of the paper?
>
